backend/requirements.txt
flask
flask-cors
groq
python-dotenv
pymupdf
python-docx
gunicorn